Added abbriviation of text
runtimetools now allows the abbriviation of words or even whole sentences, in form of removing non important characters, vowels by default. The hard keyword specifies whether all of them should be removed or if it should take a little more care of double or chars that follow eo, short words, or chars at the beggining of the word.